Moved PlayerColor to types directory
