+++ core/development/logic/pieces/Bishop.test.ts
@@ -8,7 +8,7 @@ import { Player } from '../players/Player';
 import { Bishop } from './Bishop';
 import { Pawn } from './Pawn';
 import { Position } from './PiecesUtilities';
-import { PlayerColor } from '../players/PlayerColor';
+import { PlayerColor } from '../players/types/PlayerColor';
 
 const whitePlayer = new Player(PlayerColor.WHITE);
 const blackPlayer = new Player(PlayerColor.BLACK);
