+++ core/development/logic/pieces/DoubleQueen.test.ts
@@ -9,7 +9,7 @@ import { Player } from '../players/Player';
 import { Position } from './PiecesUtilities';
 import { Pawn } from './Pawn';
 import { DoubleQueen } from './DoubleQueen';
-import { PlayerColor } from '../players/PlayerColor';
+import { PlayerColor } from '../players/types/PlayerColor';
 
 const whitePlayer = new Player(PlayerColor.WHITE);
 const blackPlayer = new Player(PlayerColor.BLACK);
