+++ core/development/logic/pieces/Golem.test.ts
@@ -8,7 +8,7 @@ import { Player } from '../players/Player';
 import { Position } from './PiecesUtilities';
 import { Pawn } from './Pawn';
 import { Golem } from './Golem';
-import { PlayerColor } from '../players/PlayerColor';
+import { PlayerColor } from '../players/types/PlayerColor';
 
 const whitePlayer = new Player(PlayerColor.WHITE);
 const blackPlayer = new Player(PlayerColor.BLACK);
