+++ core/development/logic/pieces/King.test.ts
@@ -4,7 +4,7 @@ import { onPlayerAction } from '../PieceLogic';
 import { Player } from '../players/Player';
 import { King } from './King';
 import { Position } from './PiecesUtilities';
-import { PlayerColor } from '../players/PlayerColor';
+import { PlayerColor } from '../players/types/PlayerColor';
 
 const whitePlayer = new Player(PlayerColor.WHITE);
 const blackPlayer = new Player(PlayerColor.BLACK);
