+++ core/development/logic/pieces/Knight.test.ts
@@ -5,7 +5,7 @@ import { Player } from '../players/Player';
 import { Knight } from './Knight';
 import { Position } from './PiecesUtilities';
 import { Rook } from './Rook';
-import { PlayerColor } from '../players/PlayerColor';
+import { PlayerColor } from '../players/types/PlayerColor';
 
 const whitePlayer = new Player(PlayerColor.WHITE);
 const blackPlayer = new Player(PlayerColor.BLACK);
