+++ core/development/logic/pieces/Pawn.ts
@@ -4,7 +4,7 @@ import { Player } from '../players/Player';
 import { Position } from './PiecesUtilities';
 import { game } from '../../Game';
 import { comparePositions, getPieceByPosition } from '../Utilities';
-import { PlayerColor } from '../players/PlayerColor';
+import { PlayerColor } from '../players/types/PlayerColor';
 
 export class Pawn extends Piece {
   possibleEnPassantPositions: [Position, Position] | undefined;
