+++ core/development/logic/pieces/Rook.test.ts
@@ -10,7 +10,7 @@ import { Player } from '../players/Player';
 import { Position } from './PiecesUtilities';
 import { Rook } from './Rook';
 import { Pawn } from './Pawn';
-import { PlayerColor } from '../players/PlayerColor';
+import { PlayerColor } from '../players/types/PlayerColor';
 
 const whitePlayer = new Player(PlayerColor.WHITE);
 const blackPlayer = new Player(PlayerColor.BLACK);
