+++ core/development/ui/InventoriesUI.ts
@@ -1,6 +1,6 @@
 import { INVENTORY_ELEMENT_ID, INVENTORY_WIDTH } from '../Constants';
 import { BaseItem } from '../logic/items/abstract/Item';
-import { PlayerColor } from '../logic/players/PlayerColor';
+import { PlayerColor } from '../logic/players/types/PlayerColor';
 import { initializeDraggingListeners } from './Events';
 
 const inventoryElement = document.getElementById(INVENTORY_ELEMENT_ID);
